+++ packages/auth/ng-package.json
@@ -4,6 +4,10 @@
   "deleteDestPath": true,
   "lib": {
     "flatModuleFile": "auth",
-    "entryFile": "public_api.ts"
-  }
+    "entryFile": "public_api.ts",
+    "umdModuleIds": {
+      "@delon/util": "delon.util"
+    }
+  },
+  "whitelistedNonPeerDependencies": ["tslib", "angular", "@delon/util"]
 }
