+++ packages/mock/ng-package.json
@@ -8,5 +8,6 @@
     "umdModuleIds": {
       "@delon/util": "delon.util"
     }
-  }
+  },
+  "whitelistedNonPeerDependencies": ["tslib", "angular", "@delon/util"]
 }
