+++ packages/theme/ng-package.json
@@ -14,7 +14,8 @@
       "ng-zorro-antd/i18n": "ng-zorro-antd/i18n",
       "date-fns": "DateFns",
       "@delon/acl": "delon.acl",
-      "@delon/util": "delon.util"
+      "@delon/util": "delon.util",
+      "@delon/chart": "delon.chart"
     }
   },
   "whitelistedNonPeerDependencies": [
@@ -24,6 +25,7 @@
     "ng-zorro-antd",
     "@delon/acl",
     "@delon/util",
+    "@delon/chart",
     "@ant-design/icons-angular"
   ]
 }
