+++ packages/util/src/other/deep.spec.ts
@@ -1,4 +1,4 @@
-import { copy, deepCopy, deepGet, deepMerge, deepMergeKey } from './other';
+import { deepCopy, deepGet, deepMerge, deepMergeKey } from './deep';
 
 describe('abc: utils', () => {
   describe('#deepGet', () => {
@@ -53,32 +53,6 @@ describe('abc: utils', () => {
   it('#deepCopy', () => {
     const a = { number: 1 };
     expect(deepCopy(a).number).toBe(a.number);
-  });
-
-  describe('#copy', () => {
-    it('should be copy a string', (done: () => void) => {
-      copy('test')
-        .then(() => {
-          expect(true).toBe(true);
-          done();
-        })
-        .catch(() => {
-          expect(false).toBe(true);
-          done();
-        });
-    });
-    it('[[boundary]]', (done: () => void) => {
-      spyOn(document, 'createElement').and.returnValue({ parentNode: null } as any);
-      copy('test')
-        .then(() => {
-          expect(false).toBe(true);
-          done();
-        })
-        .catch(() => {
-          expect(true).toBe(true);
-          done();
-        });
-    });
   });
 
   describe('#deepMerge', () => {
